+++ alpha/apps/kaltura/lib/batch2/kJobsManager.php
@@ -2107,17 +2107,15 @@ class kJobsManager
 		return false;
 	}
 
-	public static function addBulkUpdateUserEntryJob($partnerId, $entryId, $oldUserEntryStatus, $newUserEntryStatus)
+	public static function addUpdateUserEntriesJob($partnerId, $entryId, $oldUserEntryStatus, $newUserEntryStatus)
 	{
-		$bulkUpdateUserEntryData = new kBulkUpdateUserEntryData();
-		$bulkUpdateUserEntryData->setPartnerId($partnerId);
-		$bulkUpdateUserEntryData->setEntryId($entryId);
-		$bulkUpdateUserEntryData->setOldStatus($oldUserEntryStatus);
-		$bulkUpdateUserEntryData->setNewStatus($newUserEntryStatus);
+		$data = new kUpdateUserEntriesData();
+		$data->setOldStatus($oldUserEntryStatus);
+		$data->setNewStatus($newUserEntryStatus);
 
 		$batchJob = new BatchJob();
 		$batchJob->setPartnerId($partnerId);
 		$batchJob->setEntryId($entryId);
-		return self::addJob($batchJob, $bulkUpdateUserEntryData, BatchJobType::BULK_UPDATE_USER_ENTRY);
+		return self::addJob($batchJob, $data, BatchJobType::UPDATE_USER_ENTRIES);
 	}
 }