+++ api_v3/lib/types/batch/KalturaBatchJob.php
@@ -448,8 +448,8 @@ class KalturaBatchJob extends KalturaObject implements IFilterable
 				$this->data = new KalturaLiveEntryArchiveJobData();
 				break;
 
-			case 'kBulkUpdateUserEntryData':
-				$this->data = new KalturaBulkUpdateUserEntryData();
+			case 'kUpdateUserEntriesData':
+				$this->data = new KalturaUpdateUserEntriesData();
 				break;
 
 			default: