+++ batch/batches/RecycleBin/KAsyncUpdateUserEntriesExe.php
@@ -1,13 +1,13 @@
 <?php
 /**
- * Run KAsyncBulkUpdateUserEntry
+ * Run KAsyncUpdateUserEntries
  *
  * @package Scheduler
- * @subpackage BulkUpdateUserEntry
+ * @subpackage UpdateUserEntries
  */
 
 require_once(__DIR__ . "/../../bootstrap.php");
 
-$instance = new KAsyncBulkUpdateUserEntry();
+$instance = new KAsyncUpdateUserEntries();
 $instance->run();
 $instance->done();