+++ release-notes.md
@@ -1,11 +1,22 @@
 # Venus-22.10.0
-## Recycle bin updates big bulks of userEntries inside batch ##
+## Add user entries update batch ##
 - Issue Type: Task
 - Issue ID: PLAT-25619
 
+### Configuration ###
+Add the following to batch.ini under KAsyncCopyCaptions worker
+
+```
+[KAsyncUpdateUserEntries : JobHandlerWorker]
+id													= 810
+friendlyName										= Bulk Update User Entry
+type												= KAsyncUpdateUserEntries
+scriptPath											= batches/RecycleBin/KAsyncUpdateUserEntriesExe.php
+```
+
 ### Deployment Scripts ###
 
-    php deployment/updates/scripts/add_permissions/2025_30_12_update_user_entry_permission.php
+    php deployment/updates/scripts/add_permissions/2025_12_30_update_user_entry_permission.php
 
 
 # Venus-22.8.0